Enable SDDM autologin
We have Wayland session support on SDDM for a while now so we can
enter the Hawaii session automatically.

Some users do not understand what password they must enter, as a
matter of fact there is none but unfortunaly SDDM lacks the permission
to determine that the password was erased and keeps showing
the password box. Users just need to hit enter but they often get
confused.